Created TabLayout in NewItemActivity and created fragments for tabs
