+++ app/src/main/java/com/example/closetwear/newitem/CategoryFragment.java
@@ -1,0 +1,30 @@
+package com.example.closetwear.newitem;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.example.closetwear.R;
+
+public class CategoryFragment extends Fragment {
+
+    public CategoryFragment() {
+        // Required empty public constructor
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_category, container, false);
+    }
+}
