+++ app/src/main/java/com/example/closetwear/newitem/NameFragment.java
@@ -1,0 +1,29 @@
+package com.example.closetwear.newitem;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
+
+import com.example.closetwear.R;
+
+public class NameFragment extends Fragment {
+
+    public NameFragment() {
+        // Required empty public constructor
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_category, container, false);
+    }
+}
