+++ app/src/main/java/com/example/closetwear/newitem/SubcategoryFragment.java
@@ -1,0 +1,30 @@
+package com.example.closetwear.newitem;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
+
+import com.example.closetwear.R;
+
+
+public class SubcategoryFragment extends Fragment {
+
+    public SubcategoryFragment() {
+        // Required empty public constructor
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_subcategory, container, false);
+    }
+}
