+++ app/src/main/res/layout/activity_new_outfit.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NewOutfitActivity">
+    tools:context=".newoutfit.NewOutfitActivity">
 
     <ImageView
         android:id="@+id/outfitImg"
